Set Database and Main
