+++ main.go
@@ -2,44 +2,11 @@ package main
 
 import (
 	"Golang-Echo-MVC-Pattern/routes"
-	"database/sql"
-	"fmt"
-	_ "github.com/lib/pq"
-	"time"
+	"Golang-Echo-MVC-Pattern/settings"
 )
-
-var db *sql.DB
-
-const (
-	host     = "localhost"
-	dbname   = "postgres"
-	user     = "root"
-	password = "root"
-	schema   = "db_covid_palembang"
-)
-
-func init() {
-	var err error
-	psqlInfo := fmt.Sprintf("postgres://%s:%s@%s/%s?sslmode=disable&search_path=%s", user, password, host, dbname, schema)
-	db, err = sql.Open("postgres", psqlInfo)
-
-	if err != nil {
-		panic(err)
-	}
-
-	db.SetMaxOpenConns(80)
-	db.SetConnMaxLifetime(time.Nanosecond)
-	db.SetMaxIdleConns(0)
-
-	err = db.Ping()
-
-	if err != nil {
-		panic(err)
-	}
-}
 
 // Starting server
 func main() {
-	echo := routes.Routing.GetRoutes(routes.Routing{}, db)
+	echo := routes.Routing.GetRoutes(routes.Routing{}, settings.GetDatabase())
 	_ = echo.Start(":1103")
 }
